Chassis is now ChassisView and Chassis is a container for multiple ChassisViews
