+++ src/app/core/classes/hardware/e16.ts
@@ -1,20 +1,22 @@
 import { Container, Texture, Sprite } from 'pixi.js';
 import { DriveTray } from './drivetray';
-import { Chassis } from './chassis';
+import { Chassis} from './chassis';
+import { ChassisView } from './chassis-view';
 
 export class E16 extends Chassis {
 
   constructor(){
     super();
     this.model = "e16";
-    this.container = new PIXI.Container();
-    this.chassisPath = "assets/images/hardware/e16/e16_960w.png";
-    this.driveTrayBackgroundPath = "assets/images/hardware/e16/e16_960w_drivetray_bg.png" 
-    this.driveTrayHandlePath = "assets/images/hardware/e16/e16_960w_drivetray_handle.png"
-    this.totalDriveTrays = 16;
-    this.driveTraysOffsetY = 32;
-    this.driveTraysOffsetX = 14;
-    //this.load();
+
+    this.front = new ChassisView();
+    this.front.container = new PIXI.Container();
+    this.front.chassisPath = "assets/images/hardware/e16/e16_960w.png";
+    this.front.driveTrayBackgroundPath = "assets/images/hardware/e16/e16_960w_drivetray_bg.png" 
+    this.front.driveTrayHandlePath = "assets/images/hardware/e16/e16_960w_drivetray_handle.png"
+    this.front.totalDriveTrays = 16;
+    this.front.driveTraysOffsetY = 32;
+    this.front.driveTraysOffsetX = 14;
   }
 
 }
